Added comments and modified how the endpoints work
The main change in this update is the apiEndpoints.js and ApiTesterSiderbar.jsx. I removed the slicing and split the endpoints into groups. This will allow for easier scaling when adding/removing to existing groups. However, new groups will need to be added to the export at the bottom of the file.

The other files have just had comments added for easier comprehension.